refactor(apple): reuse more code
