+++ winio-ui-ui-kit/src/ui/mod.rs
@@ -39,9 +39,6 @@ pub use combo_box::*;
 mod list_box;
 pub use list_box::*;
 
-mod scroll_bar;
-pub use scroll_bar::*;
-
 mod scroll_view;
 pub use scroll_view::*;
